requirements.txt
# === Core Machine Learning Libraries ===
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
scipy>=1.11.0

# === Deep Learning & Transformer Models ===
torch>=2.1.0
transformers>=4.41.0
datasets>=2.19.0
tokenizers>=0.19.0

# === Configuration & Utilities ===
PyYAML>=6.0
tqdm>=4.66.0
joblib>=1.3.0

# === Visualization & Evaluation (Optional but Useful) ===
matplotlib>=3.8.0
seaborn>=0.13.0

# === Explainability (Optional Advanced Features) ===
lime>=0.2.0.1
shap>=0.44.0

# === Logging & Experiment Tracking (Optional) ===
wandb>=0.17.0

# === Code Quality / Development (Optional) ===
black>=24.0.0
flake8>=7.0.0

# === Install the Root Project as Editable Local Package ===
-e .
